+++ packages/oath/tool/dartdoc.dart
@@ -1,0 +1,7 @@
+#!/usr/bin/env dart
+
+import 'package:oath/tools/dartdoc.dart';
+
+void main() async {
+  await runDartdoc(mode: DartdocMode.preview);
+}